command methods moved into RoverVehicle class
